pyproject.toml
[tool.poetry]
name = "greenpessimisticcode"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.8"
spotipy = "^2.15.0"
pandas = "^1.1.2"
python-dotenv = "^0.14.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"